% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_downstream_lines.R
\name{get_downstream_lines}
\alias{get_downstream_lines}
\title{Get downstream lines}
\usage{
get_downstream_lines(
  strm = NA,
  jnid = "92607204bc7e4dcb939fbc57b709be0b",
  ds_sreach = 5000
)
}
\arguments{
\item{strm}{Class sf (multilinestring). Stream lines from NHN working unit (NHN_HN_NLFLOW_1)}

\item{jnid}{Character. Junction ID from NHN_HN_HYDROJUNCT_0 e.g., "92607204bc7e4dcb939fbc57b709be0b"}

\item{ds_sreach}{Numeric. Number of iterations downstream step limit.}
}
\value{
Downstream streamlines from a junction point
}
\description{
Trace downstream flow lines from a point
}
\details{
Get down streamlines traces the downstream flow path from a junction ID. Inputs include the streamline layer, jnid junction ID and ds_reach (a special parameter specifying how many steps to take).
}
\examples{
\dontrun{
filename_strm <- system.file("extdata", "NHN_HN_NLFLOW_1.gpkg", package = "canadaNHN")

# Load example data
strm <- sf::st_read(paste0(filename_strm), layer="NHN_HN_NLFLOW_1")
strm <- sf::st_zm(strm)
ds_lines <- get_downstream_lines(strm = strm,
jnid = "484667c6025244219cfa5e5ee994715b",
ds_sreach = 1000)
plot(sf::st_geometry(strm))
plot(sf::st_geometry(ds_lines), add = TRUE, col = "red")
# mapview(ds_lines)
}

}
