% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_utm_zone.R
\name{util_utm_zone}
\alias{util_utm_zone}
\title{Returns the UTM EPSG code for a given input}
\usage{
util_utm_zone(input_sf)
}
\arguments{
\item{input_sf}{Input polygons as sf}
}
\value{
The EPSG code of the local UTM zone.
}
\description{
Input should be longitude
}
